refactor: common/shared 기반 폴더 구조 재정리(#12)
