+++ src/shared/tokens/default-token.json
@@ -139,8 +139,6 @@
   },
   "$themes": [],
   "$metadata": {
-    "tokenSetOrder": [
-      "core"
-    ]
+    "tokenSetOrder": ["core"]
   }
 }
